some PyQt6 fixes

+++ kataja/saved/movables/Node.py
@@ -1170,7 +1170,7 @@ class Node(Draggable, Movable):
         self.update()
 
     def update_halo(self, color):
-        c = color.lighter(100 + (1 - ctrl.cm.background_lightness) * 120)
+        c = color.lighter(int(100 + (1 - ctrl.cm.background_lightness) * 120))
         c = ctrl.cm.transparent(c, opacity=128)
         self.halo_item.setZValue(2)
         self.halo_item.setPen(c)
